SASS-6397
