+++ src/test/resources/journeys.conf
@@ -149,7 +149,9 @@ journeys {
       individual-income-from-work,
       agent-income-from-work,
       payments-into-pensions-individual,
-      payments-into-pensions-agent
+      payments-into-pensions-agent,
+      property-pensions-investments-individual,
+      property-pensions-investments-agent
     ]
     run-if = ["journeyFlows"]
   }
