+++ src/test/scala/uk/gov/hmrc/perftests/itsass/IncomeTaxSubmissionSimulation.scala
@@ -95,6 +95,8 @@ class IncomeTaxSubmissionSimulation extends PerformanceTestRunner with Dividends
   incomeFromWorkAgent(id = "agent-income-from-work", description = "Agent Income From Work Journey")
   individualPaymentsIntoPensions(id = "payments-into-pensions-individual", description = "Individual Payments Into Pensions")
   agentPaymentsIntoPensions(id = "payments-into-pensions-agent", description = "Agent Payments Into Pensions")
+  propertyPensionsInvestmentsIndividual(id = "property-pensions-investments-individual", description = "Individual Property Pensions Investments")
+  propertyPensionsInvestmentsAgent(id = "property-pensions-investments-agent", description = "Agent Property Pensions Investments")
 
   //NRS
   nrsJourney(id = "nrs-proxy-crystallisation", description = "NRS Crystallisation Downstream Request")
